refactor: modularize face pipeline
